Add a workaround for fast test iterations
